Push state down needed for simplicity
